Exercise: Completed Assigning display properties
We’ve provided you with some divs on this page. Assign each div a display property matching the class on that div. For example: the div with a class of “inline-block” should be given a display property of “inline-block”.
Feel free to add additional div elements to see how elements with different display properties interact with each other on the page!