Exercise: Completed Sticky layout
You can use the position property to provide a simple “sticky” layout on your web pages. Using just CSS, update the header element to be fixed to the top and the footer element to be fixed to the bottom of the window.
Changing the position of these elements will also change their display properties, resulting in an unexpected side effect. It’s not necessary to pass this exercise, but can you use some other CSS properties to make the header & footer extend across the width of the page?